backend/requirements.txt
fastapi[standard]
uvicorn[standard]
motor
beanie
python-multipart
apscheduler
playwright
cryptography
python-dotenv
pydantic
pyjwt[crypto]
bcrypt
httpx
